requirements.txt
tornado==5.0.2
tornado-redis==2.4.18
tornado-smtp==0.1.1
Tornado-MySQL==0.5.1
PyMySQL==0.8.1
funcsigs==1.0.2
futures==3.1.1
geetest==3.2.1
greenlet==0.4.5
gunicorn==19.7.1
supervisor==3.3.1
pycrypto==2.6.1
redis==2.10.3
rsa==3.4.2
urllib3==1.21.1
xlrd==1.0.0
xlwt==1.2.0
xmltodict==0.11.0
aiohttp==3.4.4
asyncio-stripe==0.2.1
yunpian-sdk-python==0.0.8
python-alipay-SDK==1.7.1
qiniu==7.1.2
aiofiles==0.4.0